requirements.txt
PyMuPDF==1.23.26
Pillow==10.2.0
openai>=2.0.0
python-dotenv==1.0.1
pydantic==2.12.5
pytest==8.1.1
pytest-mock==3.12.0
fastapi==0.115.0
uvicorn==0.30.0
aiosqlite==0.20.0
pydantic-settings==2.8.1
requests==2.32.3
slowapi==0.1.9
